### 2017-06-06 #124 @ottlngr
bib2df
Provides functionality to parse BibTeX files to data.frame which can be used for analyses and visualizations using common tools like `dplyr`, `tidyr` and `ggplot2`. Changes in the data.frame can be written back to a valid BibTeX file.

```
Package: bib2df
Title: Parse a BibTeX File to a data.frame
Version: 0.2
Authors@R: c(person("Philipp", "Ottolinger", email = "philipp@ottolinger.de", role = c("aut", "cre")),
             person("Thomas", "Leeper", email = "thosjleeper@gmail.com", role = "ctb"),
             person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = "ctb"))
Description: Parse a BibTeX file to a data.frame to make it accessible for further analysis and visualization.
    plyr,
    humaniformat
    tidyr
```

https://github.com/ottlngr/bib2df

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under and why? (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.):

Data extraction, because `bib2df` makes entries of a BibTeX file accessible in the most common data structure, the data.frame.

Anyone who wants to gain insights in the characteristics and coherences of bibliographic references.

There are a few packages that can deal with BibTeX in general, like `RefManageR`#119 and others mentioned there. These packages are very close to the actual purpose of BibTeX, providing special classes and can be used as reference managing software. `bib2df` is not for managing references, but for extracting the data to a common data structure. To my knowledge, only `bib2df` offers parsing to data.frame.

- [x] Do you intend for this package to go on CRAN?  **It already is on CRAN.**

```
R CMD check results
0 errors | 0 warnings | 0 notes
```
