### 2017-06-10 #124 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @ottlngr !

Currently seeking reviewers. It's a good fit and not overlapping.

* `goodpractice::gp()` output is below. Some things for submitter and reviewers to consider.

```r
It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 91% of
    code lines are covered by test cases.

    R/bib2df_gather.R:20:NA
    R/bib2df_tidy.R:23:NA
    R/bib2df_tidy.R:30:NA
    R/df2bib.R:20:NA
    R/df2bib.R:21:NA
    ... and 3 more lines

  ✖ omit "Date" in DESCRIPTION. It is not required and it gets invalid quite
    often. A build date will be added to the package when you perform `R CMD build` on it.
  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines shorter than 80
    characters

    inst/legacy/bib2df_explore.R:2:1
    inst/legacy/bib2df_explore.R:3:1
    inst/legacy/bib2df_scrape_l.R:34:1
    inst/legacy/bib2df_scrape_l.R:39:1
    R/bib2df_gather.R:23:1
    ... and 8 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list,
    depending on the input data. Consider using vapply() instead.

    inst/legacy/bib2df_gather2_l.R:2:12
    inst/legacy/bib2df_scrape_l.R:6:9
    inst/legacy/bib2df_scrape_l.R:37:40

  ✖ not import packages as a whole, as this can cause name clashes between the
    imported packages. Instead, import only the specific functions you need.
  ✖ avoid 'T' and 'F', as they are just variables which are set to the logicals
    'TRUE' and 'FALSE' by default, but are not reserved words and hence can be overwritten
    by the user.  Hence, one should always use 'TRUE' and 'FALSE' for the logicals.

    R/bib2df_gather.R:NA:NA
    R/bib2df_tidy.R:NA:NA
    R/bib2df_tidy.R:NA:NA
```

---

Reviewers: xx xx
Due date: xxx
