### 2017-06-11 1b14d10
Remove Date from DESCRIPTION and remove legacy files
