+++ inst/legacy/bib2df_explore.R
@@ -1,6 +1,0 @@
-bib2df_explore <- function(bib) {
-  keys <- unique(str_extract(bib, "(?<=@).*(?=\\{)")[!is.na(str_extract(bib, "(?<=@).*(?=\\{)"))])
-  s <- c("Article", "article", "book", "Book", "InBook", "inbook", "InCollection", "incollection")
-  keys <- keys[keys %in% s]
-  return(keys)
-}
