+++ inst/legacy/bib2df_incollection.R
@@ -1,5 +1,0 @@
-bib2df_incollection <- function(bib) {
-  matches <- "@InCollection"
-  df <- bib2df_scrape(bib, matches)
-  return(df)
-}
