+++ inst/legacy/bib2df_inbook.R
@@ -1,5 +1,0 @@
-bib2df_inbook <- function(bib) {
-  matches <- "@InBook"
-  df <- bib2df_scrape(bib, matches)
-  return(df)
-}
