+++ DESCRIPTION
@@ -2,7 +2,6 @@ Package: bib2df
 Type: Package
 Title: Parse a BibTeX File to a data.frame
 Version: 0.2
-Date: 2017-05-21
 Authors@R: c(person("Philipp", "Ottolinger", email = "philipp@ottolinger.de", role = c("aut", "cre")),
              person("Thomas", "Leeper", email = "thosjleeper@gmail.com", role = "ctb"),
              person("Maëlle", "Salmon", email = "maelle.salmon@yahoo.se", role = "ctb"))
