### 2017-06-11 #124 @ottlngr
Updated the package regarding to the `gp()` output. This is what remains:

```
It is good practice to

  ✖ write unit tests for all functions, and all package code in general. 87% of code
    lines are covered by test cases.

    R/bib2df_gather.R:20:NA
    R/bib2df_tidy.R:25:NA
    R/bib2df_tidy.R:26:NA
    R/bib2df_tidy.R:27:NA
    R/bib2df_tidy.R:28:NA
    ... and 18 more lines

  ✖ not import packages as a whole, as this can cause name clashes between the imported
    packages. Instead, import only the specific functions you need.

```
