+++ NAMESPACE
@@ -6,7 +6,9 @@ export(bib2df_read)
 export(bib2df_tidy)
 export(df2bib)
 import(dplyr)
-import(humaniformat)
+importFrom(humaniformat,format_period)
+importFrom(humaniformat,format_reverse)
+importFrom(humaniformat,parse_names)
 importFrom(plyr,rbind.fill)
 importFrom(stats,complete.cases)
 importFrom(stringr,str_extract)
