+++ DESCRIPTION
@@ -23,3 +23,4 @@ Suggests:
     tidyr
 RoxygenNote: 6.0.1
 VignetteBuilder: knitr
+Encoding: UTF-8
