### 2017-06-15 #124 @adamhsparks
I've completed my review of _bib2df_, here are my comments and suggestions.

_bib2df_ offers functionality for working with LaTeX .bib files in R as a `data.frame` object. This functionality is well appreciated and will serve many people well. I'm actually using it at the moment in a paper that I'm writing on reproducibility in plant pathology, <https://github.com/phytopathology/Reproducible.Plant.Pathology>. The package installs quickly with no issues from CRAN or from source building the package in RStudio. The vignette and README offer good examples of how it may be used, but the package itself doesn't attempt to do more than convert a .bib file to a `data.frame()` object. I appreciate that @ottlngr has created something that is simple and does one job and does it well.

## General Comments

- Passes checks with no NOTES or WARNINGS on macOS Sierra, R 3.4.0

- Line widths are more than 80 chars in several of the .R files, using "CMD+^+A"
will reformat the code in RStudio and wrap long lines. However, some may need to be manually shortened.

- Please check spelling using `devtools::spell_check`

- Please check spelling in vignette as well. RStudio offers built-in spell-check for Rmd files, <F7>.

- Commas should always have a space after, check using `lintr::lint_package()`

- Variables and function names should all be lowercase, e.g., df2bib.R line 19

- The `details` portion of `bib2df()` documentation includes,
> Uses the internal bib2df_read(), bib2df_gather() and bib2df_tidy(). No magic, just a wrapper.

  I'm not sure that this is necessary to include for the end users, but since you have documented these internal functions you can link to the help file for them. I would suggest not including this information here.

- Following on to the previous point, I'd suggest not exposing these functions outside the package namespace and not documenting them. If I understand correctly, `bib2df_read()`, `bib2df_gather()` and `bib2df_tidy()` are all internal functions that the end-user is unlikely to interact directly with. As the package is written now, I don't see how or why a user would, so I suggest not documenting in the help files and not exporting them.

## Functions

- Using `bib2df` without specifying `separate_names =` results in a warning message. This should not happen, especially as `bib2df(path)` is the example and triggers this behaviour. I'd default to `FALSE` with the option to set to `TRUE`, but this really is your choice how you think it should be handled by default.

- Likewise, `bib2df_tidy()` exhibits the same behaviour and should not return values with a WARNING message.

- Using `bib2df_tidy()` with `separate_names` set to `TRUE` or `FALSE` results in an error message, e.g.,

  ```r
  bib <- bib2df_tidy(bib, separate_names = FALSE)
  Show Traceback

  Rerun with Debug
  Error in mutate_impl(.data, dots) :
  Evaluation error: non-character argument.
  ```
  However, as noted, I'm not sure that this function should be exported and exposed to the user?

## Documentation

- The `bib2df-package` help file could use some more formatting. You could use \code{\link{dplyr}} and the like for the other packages used. `data.frame` is not wrapped in a \code{} tag here either.

- The title field in .Rd files should not end in a period, the following files do and should be corrected,

  - bib2df.R

  - bib2df_gather.R

  - bib2df_tidy.R

  - bib2df_read.R

  - df2bib.R

- I'd change the documentation to say that this package returns a `tibble`, not a `data.frame`.

### Vignette

- The vignette is not indexed so when searching for it, `vignette(bib2df)`, the result is,

  ```r
  Warning message:
  vignette ‘bib2df’ not found
  ```

- Please make sure that the `%\VignetteIndexEntry{}` is updated, e.g.,  `%\VignetteIndexEntry{bib2df}`

- Why is it advised to download a bib file? Why not read from a URL? Some more detail in the vignette about this would be welcome. Minor point, but is a shorter URL for the example bib file possible?

- A link to the `humaniformat` package would be nice.

### NEWS.md

- Formatting of functions and R objects is inconsistent. Some use the "` `" tags,
others don't.

## Packages Used

- _plyr_ is no longer actively developed, and since _dplyr_ is already imported and has `bind_rows`, I'd suggest using this function. This also reduces dependencies by one package.

## Tests

- Tests run successfully on macOS Sierra, R 3.4.0

- Tests cover 87% bib2df_tidy.R and df2bib.R have the least coverage according to covr.io, if tests could be written to cover these more completely it would be nice.

Total time to review was about 3.5 hours.
