### 2017-06-16 #124 @ottlngr
Thanks for your extensive and detailed feedback, @adamhsparks. I will bundle your remarks in issues soon and process them as soon as possible.
