### 2017-06-22 ottlngr/bib2df#8 @ottlngr
Line widths
Avoid line widths > 80 characters.

https://github.com/ropensci/onboarding/issues/124
