### 2017-06-22 ottlngr/bib2df#9 @ottlngr
spell check
Spell check the whole package + vignette

https://github.com/ropensci/onboarding/issues/124
