### 2017-06-22 ottlngr/bib2df#10 @ottlngr
#ropensci: Formatting and Code Conventions
- [x] Avoid line widths > 80 characters
- [x] spell check on the whole package
- [x] Commas should always have a space after, use `lintr::lint_package()`
- [x] Variables and function names should all be lowercase
- [x] Improve formatting of the `bib2df-package` help file
- [x] The title field in .Rd files should not end in a period
- [x] Update and improve formatting of `NEWS.md`

https://github.com/ropensci/onboarding/issues/124#issuecomment-308617830
