### 2017-06-22 ottlngr/bib2df#11 @ottlngr
#ropensci: Do not export internal functions
These are:

* bib2df_read()
* bib2df_gather()
* bib2df_tidy()

Also, remove documentation for these functions.

https://github.com/ropensci/onboarding/issues/124#issuecomment-308617830
