### 2017-06-22 ottlngr/bib2df#12 @ottlngr
#ropensci: Avoid warning message
Using bib2df without specifying `separate_names` results in a warning message. Set default to FALSE with the option to set to TRUE.

https://github.com/ropensci/onboarding/issues/124#issuecomment-308617830
