### 2017-06-22 ottlngr/bib2df#13 @ottlngr
#ropensci: Improve the vignette
- [x] Vignettes is not indexed
- [x] Update `%\VignetteIndexEntry{}`
- [x] Link to the `humaniformat` package

https://github.com/ropensci/onboarding/issues/124#issuecomment-308617830
