### 2017-06-22 ottlngr/bib2df#14 @ottlngr
#ropensci: Remove `plyr` dependency
`plyr` is no longer actively developed, and since `dplyr` is already imported and has `bind_rows`.

https://github.com/ropensci/onboarding/issues/124#issuecomment-308617830
