### 2017-06-22 ottlngr/bib2df#15 @ottlngr
#ropensci: Tests
Increase code coverage, especially for `df2bib.R` and `bib2df_tidy.R`.

https://github.com/ropensci/onboarding/issues/124#issuecomment-308617830
