+++ R/bib2df.R
@@ -1,6 +1,6 @@
-#' @title Parse a BibTeX file to a \code{data.frame}..
-#' @description The BibTeX file is read, parsed, tidied and written to a \code{data.frame}..
-#' @details Uses the internal \code{bib2df_read()}, \code{bib2df_gather()} and \code{bib2df_tidy()}. No magic, just a wrapper.
+#' @title Parse a BibTeX file to a \code{data.frame}
+#' @description The BibTeX file is read, parsed, tidied and written to a \code{data.frame}
+#' @details For simplicity \code{bib2df()} unifies the reading, parsing and tidying of a BibTeX file while being aware of a standardized output format, different BibTeX styles and missing values in the BibTeX file.
 #' @param file character, path to a .bib file.
 #' @param separate_names logical, should authors' and editors' names be separated into first and given name?
 #' @return A \code{data.frame}.
@@ -10,7 +10,7 @@
 #' bib2df(path)
 #' @seealso \code{\link{df2bib}}
 #' @export
-bib2df <- function(file, separate_names = c(FALSE, TRUE)) {
+bib2df <- function(file, separate_names = FALSE) {
   bib <- bib2df_read(file)
   bib <- bib2df_gather(bib)
   bib <- bib2df_tidy(bib, separate_names)
