+++ R/bib2df-package.R
@@ -1,13 +1,13 @@
-#' bib2df: Parse a BibTeX file to data.frame
+#' bib2df: Parse a BibTeX file to \code{data.frame}
 #'
 #' This package provides functions to parse and write BibTeX files. BibTeX files
-#' can be parsed to data.frame to make them accessible with popular tools like
-#' dplyr, tidyr, ggplot2 and many more.
+#' can be parsed to \code{data.frame} to make them accessible with popular tools like
+#' \code{dplyr}, \code{tidyr}, \code{ggplot2} and many more.
 #'
-#' BibTeX entries represented in a data.frame can be altered in a familar way and
+#' BibTeX entries represented in a \code{data.frame} can be altered in a familiar way and
 #' written back to a valid BibTeX file.
 #'
-#' To learn more about bib2df, start with the vignettes:
+#' To learn more about \code{bib2df}, start with the vignettes:
 #'  \code{browseVignettes(package = "bib2df")}
 #' @docType package
 #' @name bib2df-package
