### 2017-06-26 ottlngr/bib2df#13 @ottlngr
https://github.com/ottlngr/bib2df/commit/2f38c206dc352bcedf34dc31c09781b16216ae4f
