### 2017-07-03 c7b5e78
Added test case
