### 2017-08-02 #124 @sckott
and a reasonable range of inputs and conditions. All tests pass on the local machine.  __Coverage is at 87% - quite good, eventually work on getting full coverage__

some bits from guidelines:
    - In releases https://github.com/ottlngr/bib2df/releases - we do want maintainers to put the relevant news items for each cran release in a git tagged release (you can link directly to issues as well, which is really helpful to give users context). Note that you can retroactively git tag commits, and add news items to releases.
    -  Example are sparse. Granted, the functions I suppose are relatively simple but should have many more covering various use cases.

2

---

Nice package!

- i wouldn't run installation of pkgs in your vignettes, both for `install.packages` and `devtools::install_github`
- for vignette: i'd suggest `warnings=FALSE` and `messages=FALSE`
- for both functions have you considered vectorizing? allowing >1 things input (either file paths or data.frames?  not sure if you have already considered and didn't make sense. Maybe could be different functions if you don't want to break functionality of current fxns.

__df2bib__

- should use `@references` to link to URLs describing .bib files
- i'd think that this function would allow the user to append or not to a file, toggling with a parameter.  e.g. `cat(..., append = TRUE)`
- you have two helper fxns inside this fxn: `capitalize` and `na_replace` - i'd put those outside of the function itself to simplify
- seems like it should fail better when passing in things that are not file paths, e.g.,

```r
df2bib(bib, 5)
#> Error in cat(paste0("@", capitalize(x$Category), "{", x$Bibtexkey, ",\n",  :
#>   invalid connection
```

__bib2df__

- `@return` should say that it returns a tibble, not a data.frame
- should note that when `separate_names = TRUE` that it gives back a list of data.frame's and when `FALSE` a list of character strings
- As other review raised above, It's good idea to support reading from a URL. Does seem to work, e.g, `bib2df("https://raw.githubusercontent.com/ottlngr/bib2df/master/inst/extdata/biblio.bib")`,
- how are malformed bib files dealt with now?  e.g., with the following malformed bib

```
@Article{Binmore2008,
  Title = {Do Conventions Need to Be Common Knowledge?},
  Author = {Binmore, Ken},
  Journal = {Topoi},
  Pages = {17--27},
  Volume = {27}
}

@Book{Osborne1994,
  Title = {A Course in Game Theory},
  Author = {Osborne, Martin J. and Rubinstein, Ariel},
  Publisher = {The MIT Press, Cambridge, Massachusetts},
  Year = {1994}
}

InCollection{BrandenburgerDekel1989,
  Title = {The Role of Common Knowledge Assumptions in Game Theory},
  Author = {Brandenburger, Adam and Dekel, Eddie},
  Booktitle = {The Economics of Missing Markets, Information and Games},
  Publisher = {Oxford University Press},
  Address = {New York},
  Chapter = {3},
  Editor = {Hahn, Frank},
  Pages = {46--61}
```

your function drops the invalid entry (last one) silently

```r
> bib2df("stuff2.bib")
# A tibble: 2 x 30
  CATEGORY   BIBTEXKEY  ADDRESS ANNOTE    AUTHOR                                               BOOKTITLE CHAPTER
     <chr>       <chr>    <chr>  <chr>    <list>                                                   <chr>   <chr>
1  ARTICLE Binmore2008     <NA>   <NA> <chr [1]>                                                    <NA>    <NA>
2     BOOK Osborne1994 New York   <NA> <chr [2]> The Economics of Missing Markets, Information and Games       3
# ... with 23 more variables: CROSSREF <chr>, EDITION <chr>, EDITOR <list>, HOWPUBLISHED <chr>, INSTITUTION <chr>,
#   JOURNAL <chr>, KEY <chr>, MONTH <chr>, NOTE <chr>, NUMBER <chr>, ORGANIZATION <chr>, PAGES <chr>, PUBLISHER <chr>,
#   SCHOOL <chr>, SERIES <chr>, TITLE <chr>, TYPE <chr>, VOLUME <chr>, YEAR <dbl>, TITLE.1 <chr>, AUTHOR.1 <chr>,
#   PUBLISHER.1 <chr>, YEAR.1 <chr>
```

while `bibtex` package errors on invalid entry

```r
bibtex::read.bib("stuff2.bib")
#> Error: Invalid bib file
```

- Seems like handling malformed values for certain things could be better. Although maybe not common, users could have e.g., a character string like `foobar` instead of `1989` for year - leading to NA's - some type checking could be done on certain inputs, or is that considered beyond scope here?
