### 2017-08-02 #124 @sckott
@ottlngr both reviews in now

let me know if you have any questions about my review or the process in general
