### 2017-08-03 #124 @ottlngr
Thanks for your review and useful comments, @sckott.

> Seems like handling malformed values for certain things could be better. Although maybe not common, users could have e.g., a character string like foobar instead of 1989 for year - leading to NA's - some type checking could be done on certain inputs, or is that considered beyond scope here?

I would not say it is beyond scope. But when creating `bib2df` my intention was not to build a tool to manage `.bib` files and especially not to build a tool that manages and "repairs" other people's `.bib` files. So I think it would be useful to have some type checking and to throw corresponding warnings. Rejecting a file or an entry because of malformed values would undermine `bib2df`'s aspiration to parse any `.bib` file. Another option would be to use character columns when non-numeric values are detected in actual numeric fields.

Will have to think about that a bit more, but there will be some kind of solutions for this as well as for the rejecting of whole entries.
