### 2017-08-04 #124 @sckott
Sounds good @ottlngr - i can see the argument against verifying bib files - just nudging a bit in direction of helping users a bit more
