### 2017-08-07 ce0a3a6
Commented out `install.packages()` and added `warning = FALSE` and `message = FALSE`
