### 2017-08-07 c17f948
Added reference to .bib format description
