### 2017-08-07 0489367
moved internal helper functions to helpers.R
