### 2017-08-07 d4a372d
moved internal helper functions to helpers.R
