### 2017-08-07 929f85b
Added type and existence checking for file paths
