### 2017-08-07 88c31f6
@return `tibble` instead of `data.frame`
