### 2017-08-07 e3786b4
Added URL as possible input type in documentation
