### 2017-08-07 bf094bd
Added check for malformed .bib entries and a respective `message()`
