### 2017-08-07 8175d9e
extended `message()`
