### 2017-08-07 c3f7e12
Added information on column type when `separate_names = TRUE`
