+++ man/bib2df.Rd
@@ -7,18 +7,20 @@
 bib2df(file, separate_names = FALSE)
 }
 \arguments{
-\item{file}{character, path to a .bib file.}
+\item{file}{character, path or URL to a .bib file.}
 
 \item{separate_names}{logical, should authors' and editors' names be separated into first and given name?}
 }
 \value{
-A \code{data.frame}.
+A \code{tibble}.
 }
 \description{
 The BibTeX file is read, parsed, tidied and written to a \code{data.frame}
 }
 \details{
 For simplicity \code{bib2df()} unifies the reading, parsing and tidying of a BibTeX file while being aware of a standardized output format, different BibTeX styles and missing values in the BibTeX file.
+
+When \code{separate_names = TRUE}, the respective columns contain a \code{data.frame} for each row. When \code{FALSE}, the respective columns contain character strings.
 }
 \examples{
 path <- system.file("extdata", "biblio.bib", package = "bib2df")
