### 2017-08-07 23b28a8
Coercion to numeric only if column `YEAR` contains no character strings
