### 2017-08-07 f49546b
Added feature to attend to existing file with `df2bib`
