### 2017-08-07 84a8eaf
Preparations for a new release
