### 2017-08-07 #124 @ottlngr
@sckott

> for both functions have you considered vectorizing? allowing >1 things input (either file paths or data.frames? not sure if you have already considered and didn't make sense. Maybe could be different functions if you don't want to break functionality of current fxns.

I think this makes sense for the `.bib` -> `data.frame` task, e.g. when read many `.bib`files from a directory. For the other way, `data.frame` -> `.bib` I do not see a use case. Even if you have read several `.bib` files with `bib2df` you can still bind these `data.frame`s together and export them back into a single `.bib` file. And the use case with exporting several `data.frame`s to several `.bib` files is some kind of "BibTeX Managing"-Task and I don't see `bib2df` pushing forward in that direction to much.

And the first case I described, reading multiple `.bib`files with `bib2df()`, is so simple using `lapply()` (could be demonstrated in the vignette) so I do not see the necessity to write a wrapper for this task. Do you think different?

> Example are sparse. Granted, the functions I suppose are relatively simple but should have many more covering various use cases.

Do you think of more examples in the documentation or in the vignette? Or both?

How does the final review process work? Do I have to provide some code that demonstrates my changes?
