+++ man/df2bib.Rd
@@ -26,7 +26,7 @@ df2bib(bib, bib2 <- tempfile())
 identical(bib, bib2df(bib2))
 }
 \references{
-\code{\link{http://www.bibtex.org/Format/}}
+\url{http://www.bibtex.org/Format/}
 }
 \seealso{
 \code{\link{bib2df}}
