### 2017-08-07 #124 @sckott
> And the first case I described, reading multiple .bibfiles with bib2df(), is so simple using lapply() (could be demonstrated in the vignette) so I do not see the necessity to write a wrapper for this task. Do you think different?

true, it is easy to do. do show how to do it

> Do you think of more examples in the documentation or in the vignette? Or both?

i think just in docs

> How does the final review process work? Do I have to provide some code that demonstrates my changes?

There isn't one way to do this. When i submit my own packages for review, I find it easiest to open an issue in the repo for each issue raised by a reviewer - and I label with `review` - then link to the issue in the review thread - You don't have to go that far: at least give a summary of what you changed/didn't change/etc. just as you would for manuscript reviews.
