### 2017-08-13 2bbbbc8
Updated examples
