+++ R/df2bib.R
@@ -7,10 +7,16 @@
 #' @author Thomas J. Leeper
 #' @references \url{http://www.bibtex.org/Format/}
 #' @examples
+#' # Read from .bib file:
 #' path <- system.file("extdata", "biblio.bib", package = "bib2df")
 #' bib <- bib2df(path)
-#' df2bib(bib, bib2 <- tempfile())
-#' identical(bib, bib2df(bib2))
+#'
+#' # Write to .bib file:
+#' bibFile <- tempfile()
+#' df2bib(bib, bibFile)
+#'
+#' # Use `append = TRUE` to add lines to an existing .bib file:
+#' df2bib(bib, bibFile, append = TRUE)
 #' @seealso \code{\link{bib2df}}
 #' @export
 df2bib <- function(x, file, append = FALSE) {
