### 2017-08-13 2684799
Updated examples
