### 2017-08-13 b8dd36a
Adding tests for helper functions
