### 2017-08-13 0ba726c
test df2bib() error messages
