### 2017-08-13 7aa2e82
Updated vignette
