### 2017-08-30 d8b8003
detection and error handling of invalid file paths
