### 2017-08-30 535d916
invalid url detection and error handling
