### 2017-08-30 d36bddb
detect invalid urls
