### 2017-08-30 efd7867
handle invalid urls
