+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 export(bib2df)
 export(df2bib)
+import(RCurl)
 import(dplyr)
 importFrom(dplyr,bind_rows)
 importFrom(humaniformat,format_period)
