### 2017-08-30 #124 @ottlngr
Sorry for the delay! Here is what I achieved based on your reviews:

- Improved formatting, respected code conventions and spell checked the package
- Avoided unnecessary dependencies (`plyr`)
- Hid internal functions from the user by not exporting them
- Moved helper functions to `helpers.R`
- Avoided warning messages and improved error handling
- Improved the vignette
- Linked to the `humaniformat` package and to a BibTex reference
- `df2bib2()` now supports appending to existent files
- Improved error messages when reading from or writing to invalid file paths or URLs
- Added a message when a malformed .bib file is recognized as such while reading

I tried to document each step in a single commit with a somewhat meaningful commit message: https://github.com/ottlngr/bib2df/commits/master

If you have any questions or want some kind of demonstration of specific (new) features just let me know.

Thanks to @adamhsparks and @sckott again your helpful and detailed reviews!
