### 2017-08-30 #124 @adamhsparks
Thanks @ottlngr, this all sounds great. I'll have a look at what you've done this weekend.
