### 2017-09-02 #124 @adamhsparks
Thanks for your efforts to address my comments, @ottlngr. I know that I'm being picky here, but I think that there are a number issues to be addressed first before I can check off on accepting this package.

## Comments on code and documentation

1. Due to issues with _covr_ on my machine, I can't use the _goodpractice_ package, but I can still run `lintr::lint_package()`, which raises several easily correctable flags. Correct any flags that this raises. Nothing major, but it will help clean up the package code.

2. Check spelling using `devtools::spell_check()`, I found at least one word that really was misspelled and not because it was a name or American (not Australian) English.

3. The bib2df.R file has extra "."s on lines 1 and 2 and bib2df_tidy.R has an extra "." on line 1.

4. In the examples for `bib2df_tidy()`, I'm getting error messages? This only occurs if I don't assign the output of `bib2df_tidy()` to an object in my R session.
```r
Warning messages:
1: In if (separate_names) { :
  the condition has length > 1 and only the first element will be used
2: In if (separate_names) { :
  the condition has length > 1 and only the first element will be used
```

5. The help file for `bib2df()` says:
`Uses the internal bib2df_read(), bib2df_gather() and bib2df_tidy(). No magic, just a wrapper.` I'm not clear on this. This indicates that these functions are internal, so I shouldn't ever need to interact with them. However, they're documented and exposed to the end-user and see previous, point 5, about the error when using `bib2df_tidy()`. But then they are not documented in the vingette.

6. I'm not clear from the documentation about whether this package is centred around `tibble()` or `data.frame()`? The vignette suggests that it's a normal `data.frame()` but the functions appear to return `tibble()`?

## Comments on the vignette

1. Check spelling in the vignette using the built-in spellcheck in RStudio for Rmd files.

2. You switch between "BibTex" and "BibTeX" in the vignette. You should use "BibTeX".

3. In chunk 1, you should probably use `{r, eval=FALSE}` to avoid reinstalling the package everytime the vignette is built. This will also squash the
```r
## Installing package into '/usr/local/lib/R/3.4/site-library'
## (as 'lib' is unspecified)
```
output.

4. In the vignette, you say that the packages in the _tidyverse_ use the `data.frame` as a data structure. Maybe I'm nitpicking because a `tibble` is a form of a `data.frame`, but I tend to think that the `tidyverse` uses `tibbles`. I don't see how the _tidyverse_ is necessary here. You can simply state that R uses `data.frame` objects as one of it's basic data structures unless there is a compelling reason to mention the _tidyverse_. I'm not against the _tidyverse_ _per se_, but I don't see what it adds to the documentation here to call it out specifically. Unless you are returning a `tibble`, in which case you should highlight that you're using the latest and greatest version of a `data.frame` in R that is also used in the `tidyverse`.

5. In the vignette you state:
`Since all the BibTeX entries are represented by rows in a data.frame, all  entries can be altered just like one would alter the value of a certain cell in a data.frame.`
Well, yes. But you seem to suggest that a `data.frame` from `bib2df` is not a normal `data.frame` object. I'd suggest rephrasing this to avoid confusion.
