+++ NAMESPACE
@@ -2,12 +2,15 @@
 
 export(bib2df)
 export(df2bib)
-import(dplyr)
 importFrom(RCurl,url.exists)
+importFrom(dplyr,as_data_frame)
 importFrom(dplyr,bind_rows)
+importFrom(dplyr,mutate)
+importFrom(dplyr,select)
 importFrom(humaniformat,format_period)
 importFrom(humaniformat,format_reverse)
 importFrom(humaniformat,parse_names)
+importFrom(magrittr,"%>%")
 importFrom(stats,complete.cases)
 importFrom(stringr,str_extract)
 importFrom(stringr,str_match)
