+++ NAMESPACE
@@ -3,6 +3,7 @@
 export(bib2df)
 export(df2bib)
 importFrom(RCurl,url.exists)
+importFrom(dplyr,"%>%")
 importFrom(dplyr,as_data_frame)
 importFrom(dplyr,bind_rows)
 importFrom(dplyr,mutate)
@@ -10,7 +11,6 @@ importFrom(dplyr,select)
 importFrom(humaniformat,format_period)
 importFrom(humaniformat,format_reverse)
 importFrom(humaniformat,parse_names)
-importFrom(magrittr,"%>%")
 importFrom(stats,complete.cases)
 importFrom(stringr,str_extract)
 importFrom(stringr,str_match)
