### 2017-09-03 f426d5d
Consistent usage of `tibble` instead of `data.frame`
