+++ vignettes/bib2df.Rmd
@@ -24,9 +24,9 @@ The resulting BibTeX file can tell much about the work you use it for, may it be
 
 ## Why bib2df?
 
-As mentioned above, BibTeX represents the entries as lists of named fields, some kind of similar to the `JSON` format. If you want to gain insights from your BibTex file using R, you will have to transform the data to fit into a more usual data structure. This data structure, speaking of R and especially the great packages of the `tidyverse` like `dplyr` and `ggplot2`, is the `data.frame`. `bib2df` does exactly this: It takes a BibTeX file and parses it into a `data.frame` so you can work with your bibliographic data just the way you do with other data.
+As mentioned above, BibTeX represents the entries as lists of named fields, some kind of similar to the `JSON` format. If you want to gain insights from your BibTex file using R, you will have to transform the data to fit into a more usual data structure. Such a data structure, speaking of R,  is the `tibble`. `bib2df` does exactly this: It takes a BibTeX file and parses it into a `tibble` so you can work with your bibliographic data just the way you do with other data.
 
-Given this `data.frame` you can manipulate entries in a familiar way and write the updated references back to a valid BibTex file. 
+Given this `tibble` you can manipulate entries in a familiar way and write the updated references back to a valid BibTex file. 
 
 ## How to use
 
@@ -41,7 +41,7 @@ library(bib2df)
 
 ### Parse the BibTeX file
 
-To parse a BibTeX file to a `data.frame` you may want to use the function `bib2df()`. The first argument of `bib2df()` is the path to the file you want to parse. If you obtain your file from an URL, use the URL as the path. 
+To parse a BibTeX file to a `tibble` you may want to use the function `bib2df()`. The first argument of `bib2df()` is the path to the file you want to parse. If you obtain your file from an URL, use the URL as the path. 
 
 ```{r}
 url <- "https://gist.githubusercontent.com/ottlngr/d709ab6c7de08d133435cb8c77699914/raw/153f2ec0be6d36c7dd9fa3389cc87f54ecf4da04/LiteratureOnCommonKnowledgeInGameTheory.bib"
@@ -50,7 +50,7 @@ df <- bib2df(url)
 df
 ```
 
-`bib2df()` returns a `data.frame` with each row representing one entry of the initial BibTeX file while the columns hold the data originally stored in the named fields. If a field was not present in a particular entry, the respective column gets the value `NA`. As some works can be the work of multiple authors or editors, these fields are converted to a list to avoid having the names of multiple persons concatenated to a single character string:
+`bib2df()` returns a `tibble` with each row representing one entry of the initial BibTeX file while the columns hold the data originally stored in the named fields. If a field was not present in a particular entry, the respective column gets the value `NA`. As some works can be the work of multiple authors or editors, these fields are converted to a list to avoid having the names of multiple persons concatenated to a single character string:
 
 ```{r}
 head(df$AUTHOR)
@@ -65,7 +65,7 @@ head(df$AUTHOR)
 
 ### Parsing multiple BibTex files
 
-Multiple BibTex files can be parsed using `lapply()`. The paths to the BibTex files must be stored in a vector. Using this vector to call `bib2df()` within `lapply()` results in a list of `data.frames`, which can be bound, e.g. using `bind_rows()`:
+Multiple BibTex files can be parsed using `lapply()`. The paths to the BibTex files must be stored in a vector. Using this vector to call `bib2df()` within `lapply()` results in a list of `tibble`, which can be bound, e.g. using `bind_rows()`:
 
 ```{r}
[…]
