### 2017-09-03 b4fbe67
Consistent usage of `tibble` instead of `data.frame`
