### 2017-09-03 312a6b3
using httr::http_error() instead of RCurl::url.exists()
