+++ R/bib2df.R
@@ -5,7 +5,7 @@
 #' @param file character, path or URL to a .bib file.
 #' @param separate_names logical, should authors' and editors' names be separated into first and given name?
 #' @return A \code{tibble}.
-#' @importFrom RCurl url.exists
+#' @importFrom httr http_error
 #' @author Philipp Ottolinger
 #' @examples
 #' # Read from .bib file:
@@ -24,7 +24,7 @@ bib2df <- function(file, separate_names = FALSE) {
     stop("Invalid file path: Non-character supplied.", call. = FALSE)
   }
   if(grepl("http://|https://|www.", file)) {
-    if(!url.exists(file)) {
+    if(http_error(file)) {
       stop("Invalid URL: File is not readable.", call. = FALSE)
     }
   } else {
