+++ NAMESPACE
@@ -2,12 +2,12 @@
 
 export(bib2df)
 export(df2bib)
-importFrom(RCurl,url.exists)
 importFrom(dplyr,"%>%")
 importFrom(dplyr,as_data_frame)
 importFrom(dplyr,bind_rows)
 importFrom(dplyr,mutate)
 importFrom(dplyr,select)
+importFrom(httr,http_error)
 importFrom(humaniformat,format_period)
 importFrom(humaniformat,format_reverse)
 importFrom(humaniformat,parse_names)
