### 2017-09-03 befd7dc
spell checked the vignette
