### 2017-09-03 6f65e34
Removed confusing statement in vignette
