### 2017-09-03 afd0aff
changes due to lintr::lint_package()
