### 2017-09-03 #124 @ottlngr
Thanks, @adamhsparks for the feedback, especially for the hint to use `httr::http_error()`!

`goodpractice::gp()` still returns missing unit tests, but I'm working on that.

I agreed to all your comments, see the commits below for details:

https://github.com/ottlngr/bib2df/commit/d21428e969ad35893e11ddea939bee9c00b1ec02
https://github.com/ottlngr/bib2df/commit/eb02636712a0d9200f4dc0e5f50581ea5200a5d3
https://github.com/ottlngr/bib2df/commit/acfcd82e18d9366f52a98bc65665745a16294142
https://github.com/ottlngr/bib2df/commit/f426d5deac08543b6550d232ade093ea7c69c29c
https://github.com/ottlngr/bib2df/commit/b4fbe674bcd201d4bcbf0d7c2e12ba7ea733ba94
https://github.com/ottlngr/bib2df/commit/312a6b362911aa8868ccc35c8bc0e61dbb6c8f94
https://github.com/ottlngr/bib2df/commit/befd7dceaaf0ff9b5354c71db3c3f54f168970ee
https://github.com/ottlngr/bib2df/commit/6f65e34c095bf7ab79042486526899984e3a6bbc
https://github.com/ottlngr/bib2df/commit/afd0aff012886d9694d43dca751b5aea3c7bdf0b
