### 2017-09-03 cfc8426
added some unit tests
