### 2017-09-03 #124 @adamhsparks
Thanks @ottlngr.

Line 65 of README.Rmd, please correct the spelling of "Community"

`## Comunity Guidelines`

Aside from that, looks good to me.
