+++ README.Rmd
@@ -62,6 +62,6 @@ Version 0.2 is now available on **CRAN**:
 install.packages("bib2df")
 ```
 
-## Comunity Guidelines
+## Community Guidelines
 
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
