+++ README.md
@@ -2,8 +2,8 @@
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 [![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/bib2df)](https://cran.r-project.org/package=bib2df)[![Travis-CI Build Status](https://travis-ci.org/ottlngr/bib2df.svg?branch=master)](https://travis-ci.org/ottlngr/bib2df)[![](http://cranlogs.r-pkg.org/badges/bib2df)](http://cran.rstudio.com/web/packages/bib2df/index.html)[![codecov](https://codecov.io/gh/ottlngr/bib2df/branch/master/graph/badge.svg)](https://codecov.io/gh/ottlngr/bib2df)
 
-`bib2df` - Parse a BibTeX file to a data.frame
-----------------------------------------------
+`bib2df` - Parse a BibTeX file to a tibble
+------------------------------------------
 
 Everyone writing reports and articles with LaTeX has probably used BibTeX before. BibTeX is the de facto standard for reference management and grounds its functionality on a list of references stored in local text file. Depending on the reference type, several fields are necessary to define a reference properly. An exemplary BibTeX entry looks as follows:
 
@@ -17,17 +17,17 @@ Everyone writing reports and articles with LaTeX has probably used BibTeX before
       Volume = {27}
     }
 
-Parse the BibTeX file to a data.frame
--------------------------------------
+Parse the BibTeX file to a tibble
+---------------------------------
 
-The BibTeX format is not convenient for any kind of analysis or visualization. Many R applications require a `data.frame` and `bib2df` offers a straightforward framework to parse a BibTeX file to a `data.frame`.
+The BibTeX format is not convenient for any kind of analysis or visualization. Many R applications require a `data.frame` (or `tibble`) and `bib2df` offers a straightforward framework to parse a BibTeX file to a `tibble`.
 
 ``` r
 library(bib2df)
 path <- system.file("extdata", "biblio.bib", package = "bib2df")
 bib <- bib2df(path)
 bib
-#> # A tibble: 3 × 26
+#> # A tibble: 3 x 26
 #>       CATEGORY              BIBTEXKEY  ADDRESS ANNOTE    AUTHOR
 #>          <chr>                  <chr>    <chr>  <chr>    <list>
 #> 1      ARTICLE            Binmore2008     <NA>   <NA> <chr [1]>
@@ -41,7 +41,7 @@ bib
 #> #   YEAR <dbl>
 ```
 
-The `df2bib()` function makes it possible to write this `data.frame` back to disk, enabling programmatic manipulation of a .bib file.
+The `df2bib()` function makes it possible to write this `tibble` back to disk, enabling programmatic manipulation of a .bib file.
 
 Installation
 ------------
@@ -54,7 +54,7 @@ Version 0.2 is now available on **CRAN**:
 
     install.packages("bib2df")
 
-Comunity Guidelines
--------------------
+Community Guidelines
+--------------------
 
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
