### 2017-09-04 #124 @ottlngr
Sorry for the typo - just noticed I had not knitted the .Rmd but now all changes are in place: https://github.com/ottlngr/bib2df/commit/4ca592b6ecd2a9ef02fbd5561ddd994b641e1531

Thanks for the approval, all in all a very helpful and educational process!
