### 2017-09-04 2fef769
added rOpenSci status badge
