### 2017-09-04 1ea51b4
added rOpenSci status badge
