### 2017-09-05 #124 @sckott
@ottlngr

* instead of commenting out `install.packages` in your vignette, i meant to say use `{r eval=FALSE}` in the knit chunk - so for the reader they can run it easily, but it's not actually run during the build process

Looks good otherwise.
