### 2017-09-05 #124 @adamhsparks
I agree with @sckott on this point. You can silently load _bib2df_ elsewhere without echoing the code so that the library loads to execute the examples that follow.
