### 2017-09-06 8812f1a
removed package installation in vignette and put `library()` in {r eval=FALSE}
