### 2017-09-06 aeb2a15
removed `install.packages()` from vignette and loaded the package silently
