### 2017-09-06 #124 @ottlngr
Thanks for clarification. I had a look at other packages' vignettes - I removed the `install.packages()` and load the package with `echo=FALSE,message=FALSE`. I guess you're ok with that?
