### 2017-09-06 #124 @sckott
@ottlngr sorry, i meant `eval = FALSE` - still want to echo it so the user sees that they probably need to isntall the pkg brefore using it 😄
