### 2017-09-06 #124 @ottlngr
I'm totally confused now about that simple task :joy:

I had a look at other vignettes (like https://github.com/ropensci/tabulizer/blob/master/vignettes/tabulizer.Rmd), many do not put `install.packages()` inside the vignette. But if one does, `eval = FALSE` is the chunk option of choice, @sckott? This makes sense by all means.

And the `library()` command without any chunk option then?
