### 2017-09-06 #124 @sckott
Yes, that's right, `install.packages`  in a chunk with `eval = FALSE`, and `library` in a chunk that does run (`eval = TRUE` the default)
