### 2017-09-06 d401345
restructured code chunks and chunk options
