+++ vignettes/bib2df.Rmd
@@ -34,11 +34,13 @@ Given this `tibble` you can manipulate entries in a familiar way and write the u
 
 To parse a BibTeX file to a `tibble` you may want to use the function `bib2df()`. The first argument of `bib2df()` is the path to the file you want to parse. If you obtain your file from an URL, use the URL as the path. 
 
-```{r, echo=FALSE,message=FALSE}
-library(bib2df)
+```{r, eval = FALSE}
+install.packages("bib2df")
 ```
 
 ```{r}
+library(bib2df)
+
 url <- "https://gist.githubusercontent.com/ottlngr/d709ab6c7de08d133435cb8c77699914/raw/153f2ec0be6d36c7dd9fa3389cc87f54ecf4da04/LiteratureOnCommonKnowledgeInGameTheory.bib"
 
 df <- bib2df(url)
