### 2017-09-06 #124 @ottlngr
:sweat_smile:

https://github.com/ottlngr/bib2df/commit/d4013453d4868c11acf291d2d977a9ce5a4d83ba
