### 2017-09-15 #124 @ottlngr
Thanks for guiding me through the whole process, @adamhsparks and @sckott !

I will transfer the repo to rOpenSci this evening (CEST). As soon as the transfer is completed, I will make the needed changes.

I'm definitely interested in writing a post for https://ropensci.org/tech-notes/, and maybe even a long-form post for https://ropensci.org/blog - depends on the existence of a suitable narrative and enough spare time to not delay the publication.
